+++ packages/javascript/lib/node/os/getCpu.ts
@@ -1,0 +1,5 @@
+import { cpus } from 'os';
+
+export function getCpu(): string {
+	return `${cpus()?.[0]?.model || 'Unknown'}`;
+}